simplify contexts and move AlertMessage context outside the package
